Update packages and add more packets

+++ Zeltlager/Zeltlager/Member.cs
@@ -8,7 +8,7 @@ namespace Zeltlager
 {
 	public class Member : IComparable<Member>
 	{
-		public uint Id { get; set; }
+		public ushort Id { get; set; }
 		public string Name { get; set; }
 		public Tent Tent { get; set; }
 		public bool Supervisor { get; set; }
@@ -19,7 +19,7 @@ namespace Zeltlager
 		{
 		}
 
-		public Member(uint id, string name, Tent tent, bool supervisor)
+		public Member(ushort id, string name, Tent tent, bool supervisor)
 		{
 			Id = id;
 			Name = name;
